refactor: Image 도메인 수정, Post엔티티 수정
