+++ src/main/java/backend/airo/application/post/usecase/PostReadUseCase.java
@@ -17,11 +17,10 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Page;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-
+import static backend.airo.domain.post.exception.PostErrorCode.POST_ACCESS_DENIED;
 import java.util.ArrayList;
 import java.util.List;
 
-import static backend.airo.domain.post.exception.PostErrorCode.POST_ACCESS_DENIED;
 
 @Slf4j
 @Service
@@ -34,20 +33,13 @@ public class PostReadUseCase {
     private final GetImageQueryService getImageQueryService;
     private final GetPostListQueryService getPostListQueryService;
 
-    // ===== 게시물 조회 =====
-
-    /**
-     * 게시물 단건 조회
-     */
     public PostDetailResponse getPostById(Long postId, Long requesterId) {
         log.debug("게시물 조회: id={}, requesterId={}", postId, requesterId);
 
         Post post = getPostQueryService.handle(postId);
 
-        // 접근 권한 검증
         validatePostAccess(post, requesterId);
 
-        // 조회수 증가
         if(!isPostOwner(post, requesterId)) {
             post.incrementViewCount();
         }
@@ -62,10 +54,6 @@ public class PostReadUseCase {
     }
 
 
-
-    /**
-     * 최근 게시물 목록 조회
-     */
     public Page<Post> getRecentPostList(PostListRequest request) {
         log.debug("최근 게시물 목록 조회: page={}, size={}, userId={}",
                 request.page(), request.size(), request.userId());
@@ -75,17 +63,14 @@ public class PostReadUseCase {
     }
 
 
-
     private AuthorInfo getAuthorInfo(Long autherId) {
 
         User author = getUserQueryService.handle(autherId);
 
-        // 작성자 정보 반환
         return new AuthorInfo(author.getId(), author.getName(), author.getProfileImageUrl());
     }
-    /**
-     * 게시물 접근 권한 검증
-     */
+
+
     private void validatePostAccess(Post post, Long requesterId) {
         if ( isPostOwner(post, requesterId)) {
             throw new PostAccessDeniedException(post.getId(), requesterId, POST_ACCESS_DENIED);
@@ -97,9 +82,6 @@ public class PostReadUseCase {
     }
 
 
-    /**
-     * 게시물 소유자 확인
-     */
     private boolean isPostOwner(Post post, Long userId) {
         return userId != null && userId.equals(post.getUserId());
     }
