--- src/main/java/backend/airo/domain/post/PostTag.java
@@ -1,4 +1,0 @@
-package backend.airo.domain.post;
-
-public class PostTag {
-}
