+++ src/test/java/backend/airo/application/image/usecase/ImageCreateUseCaseTest.java
@@ -27,7 +27,7 @@ import static org.mockito.ArgumentMatchers.*;
 import static org.mockito.Mockito.*;
 
 @ExtendWith(MockitoExtension.class)
-class ImageUseCaseTest {
+class ImageCreateUseCaseTest {
 
     @Mock
     private ImageRepository imageRepository;
